requirements.txt
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
drf-yasg==1.21.8
gunicorn==23.0.0
pillow==11.3.0
psycopg2==2.9.10
python-decouple==3.8
django-currentuser==0.9.0
